Multi-Poller changes

+++ trunk/cacti/poller.php
@@ -42,48 +42,88 @@ include_once($config["base_path"] . "/lib/data_query.php");
 include_once($config["base_path"] . "/lib/graph_export.php");
 include_once($config["base_path"] . "/lib/rrd.php");
 
-/* Record Start Time */
+/* determine the poller_id if specified */
+$poller_id = 0;
+if ( $_SERVER["argc"] == 2 ) {
+	$poller_id = $_SERVER["argv"][1];
+	if (!is_numeric($poller_id)) {
+		cacti_log("ERROR: The Poller ID is not numeric", true, "POLLER");
+		exit -1;
+	}
+}
+
+/* let PHP run just as long as it has to */
+ini_set("max_execution_time", "0");
+
+/* record start time */
 list($micro,$seconds) = split(" ", microtime());
 $start = $seconds + $micro;
 
-/* Let PHP Run Just as Long as It Has To */
-ini_set("max_execution_time", "0");
+/* default the number of pollers to 0 */
+$num_pollers = 0;
 
-/* Get number of polling items from the database */
-$num_polling_items = db_fetch_cell("select count(*) from poller_item");
-$polling_hosts = array_merge(array(0 => array("id" => "0")), db_fetch_assoc("select id from host where disabled = '' order by id"));
+/* poller_id 0 tasks only */
+if ($poller_id == 0) {
+	/* get total number of polling items from the database for the specified poller */
+	$num_polling_items = db_fetch_cell("SELECT count(*) FROM poller_item WHERE poller_id=0");
+	$polling_hosts = array_merge(array(0 => array("id" => "0")), db_fetch_assoc("SELECT id FROM host WHERE (disabled = '' AND poller_id=0) ORDER BY id"));
 
-/* Retreive the number of concurrent process settings */
+	/* get total number of polling items from the database for all pollers */
+	$all_num_polling_items = db_fetch_cell("SELECT count(*) FROM poller_item WHERE poller_id=0");
+	$all_polling_hosts = array_merge(array(0 => array("id" => "0")), db_fetch_assoc("select id from host where disabled = '' ORDER BY id"));
+
+	/* get the number of active pollers */
+	$pollers = db_fetch_assoc("SELECT * FROM poller WHERE active = 'on'");
+	$num_pollers = sizeof($pollers) + 1;
+
+	/* update web paths for the poller */
+	db_execute("REPLACE INTO settings (name,value) VALUES ('path_webroot','" . addslashes(($config["cacti_server_os"] == "win32") ? strtolower(substr(dirname(__FILE__), 0, 1)) . substr(dirname(__FILE__), 1) : dirname(__FILE__)) . "')");
+
+	/* initialize poller_time and poller_output tables */
+	db_execute("TRUNCATE TABLE poller_time");
+
+	/* open a pipe to rrdtool for writing */
+	$rrd_processes = read_config_option("concurrent_rrd_processes");
+	$rrdtool_pipe = rrd_init($rrd_processes);
+
+	/* insert the current date/time for graphs */
+	db_execute("REPLACE into settings (name,value) values ('date',NOW())");
+
+	/* allow remote pollers to start */
+	db_execute("UPDATE poller SET run_state='Ready' where active='on'");
+} else {
+	/* get total number of polling items from the database for the specified poller */
+	$num_polling_items = db_fetch_cell("SELECT count(*) FROM poller_item WHERE poller_id='" . $poller_id . "'");
+	$polling_hosts = db_fetch_assoc("SELECT id FROM host WHERE (disabled = '' and poller_id = '" . $poller_id . "') ORDER BY id");
+
+	db_execute("UPDATE poller SET run_state='Running' where id=" . $poller_id);
+}
+
+/* retreive the number of concurrent process settings */
 $concurrent_processes = read_config_option("concurrent_processes");
 
-/* Initialize counters for script file handling */
+/* initialize counters for script file handling */
 $host_count = 1;
 
-/* Initialize file creation flags */
+/* initialize file creation flags */
 $change_files = False;
 
-/* Initialize file and host count pointers */
+/* initialize file and host count pointers */
 $process_file_number = 0;
 $first_host = 0;
 $last_host = 0;
 
-/* Update web paths for the poller */
-db_execute("replace into settings (name,value) values ('path_webroot','" . addslashes(($config["cacti_server_os"] == "win32") ? strtolower(substr(dirname(__FILE__), 0, 1)) . substr(dirname(__FILE__), 1) : dirname(__FILE__)) . "')");
-
-// Obtain some defaults from the database
-$poller = read_config_option("poller_type");
+/* obtain some defaults from the database */
+$poller_type = read_config_option("poller_type");
 $max_threads = read_config_option("max_threads");
 
-// Initialize poller_time and poller_output tables
-db_execute("truncate table poller_time");
-
-// Enter Mainline Processing
-if (($num_polling_items > 0) && (read_config_option("poller_enabled") == "on")) {
+/* enter mainline processing */
+if ((($num_polling_items > 0) || ($num_pollers > 1)) && (read_config_option("poller_enabled") == "on")) {
 	/* Determine the number of hosts to process per file */
 	$hosts_per_file = ceil(sizeof($polling_hosts) / $concurrent_processes );
 
 	/* Determine Command Name */
-	if (($config["cacti_server_os"] == "unix") and ($poller == "2")) {
+	if (($config["cacti_server_os"] == "unix") and ($poller_type == "2")) {
 		$command_string = read_config_option("path_cactid");
 		$extra_args = "";
 		$method = "cactid";
@@ -92,7 +132,7 @@ if (($num_polling_items > 0) && (read_config_option("poller_enabled") == "on")) 
 		$command_string = read_config_option("path_php_binary");
 		$extra_args = "-q " . $config["base_path"] . "/cmd.php";
 		$method = "cmd.php";
-	}else if ($poller == "2") {
+	}else if ($poller_type == "2") {
 		$command_string = read_config_option("path_cactid");
 		$extra_args = "";
 		$method = "cactid";
@@ -103,7 +143,7 @@ if (($num_polling_items > 0) && (read_config_option("poller_enabled") == "on")) 
 		$method = "cmd.php";
 	}
 
-	/* Populate each execution file with appropriate information */
+	/* execute each process with the host list */
 	foreach ($polling_hosts as $item) {
 		if ($host_count == 1) {
 			$first_host = $item["id"];
@@ -117,71 +157,107 @@ if (($num_polling_items > 0) && (read_config_option("poller_enabled") == "on")) 
 		$host_count ++;
 
 		if ($change_files) {
-			exec_background($command_string, "$extra_args $first_host $last_host");
+			exec_background($command_string, "$extra_args -f=$first_host -l=$last_host -p=$poller_id");
 
 			$host_count = 1;
 			$change_files = False;
 			$process_file_number++;
 			$first_host = 0;
 			$last_host = 0;
-		} /* End change_files */
-	} /* End For Each */
+		} /* end change_files */
+	} /* end for each */
 
+	/* execute the last process if present */
 	if ($host_count > 1) {
 		$last_host = $item["id"];
 
-		exec_background($command_string, "$extra_args $first_host $last_host");
+		exec_background($command_string, "$extra_args -f=$first_host -l=$last_host -p=$poller_id");
 		$process_file_number++;
 	}
 
-	/* insert the current date/time for graphs */
-	db_execute("replace into settings (name,value) values ('date',NOW())");
-
-	if ($poller == "1") {
+	if ($poller_type == "1") {
 		$max_threads = "N/A";
 	}
 
-	/* open a pipe to rrdtool for writing */
-	$rrd_processes = read_config_option("concurrent_rrd_processes");
-	$rrdtool_pipe = rrd_init($rrd_processes);
-
 	$loop_count = 0;
 	while (1) {
-		$polling_items = db_fetch_assoc("select poller_id,end_time from poller_time where poller_id = 0");
+		$polling_items = db_fetch_assoc("SELECT poller_id,end_time FROM poller_time WHERE poller_id = " . $poller_id);
 
 		if (sizeof($polling_items) == $process_file_number) {
-			process_poller_output($rrdtool_pipe);
+			/* set poller status complete */
+			if ($poller_id != 0) {
+				db_execute("UPDATE poller SET run_state='Complete' WHERE id=" . $poller_id);
+				break;
+			} else {
+				/* process RRD output if any exists */
+				process_poller_output($rrdtool_pipe);
 
-			/* take time and log performance data */
-			list($micro,$seconds) = split(" ", microtime());
-			$end = $seconds + $micro;
+				/* wait for other pollers to finish */
+				$total_pollers = sizeof(db_fetch_assoc("SELECT * FROM poller WHERE active='on'"))+1;
+				while (1) {
+					$active_pollers = sizeof(db_fetch_assoc("SELECT * FROM poller WHERE (active='on' AND run_state != 'Complete')"));
+					if ($active_pollers == 0) {
+						process_poller_output($rrdtool_pipe);
 
-			cacti_log(sprintf("STATS: " .
-				"Time: %01.4f s, " .
-				"Method: %s, " .
-				"Processes: %s, " .
-				"Threads: %s, " .
-				"Hosts: %s, " .
-				"Hosts/Process: %s",
-				round($end-$start,4),
-				$method,
-				$concurrent_processes,
-				$max_threads,
-				sizeof($polling_hosts),
-				$hosts_per_file),true,"SYSTEM");
+						break;
+					} else {
+						process_poller_output($rrdtool_pipe);
 
-			break;
+						if (read_config_option("log_verbosity") >= POLLER_VERBOSITY_MEDIUM) {
+							print "Poller 0 Complete.  Waiting on " . ($total_pollers - $active_pollers) . " Pollers to complete.\n";
+						}
+
+						if (($start + MAX_POLLER_RUNTIME) < time()) {
+							/* close rrdtool if poller is 0 */
+							if ($poller_id == 0) { rrd_close($rrdtool_pipe); }
+							cacti_log("ERROR: Maximum runtime of " . MAX_POLLER_RUNTIME . " seconds exceeded for Poller_ID " . $poller_id . " - Exiting.", true, "POLLER");
+							exit;
+						}
+					}
+
+					/* give pollers another second to end */
+					sleep(1);
+				}
+
+				/* take time and log performance data */
+				list($micro,$seconds) = split(" ", microtime());
+				$end = $seconds + $micro;
+
+				cacti_log(sprintf("STATS: " .
+					"Time: %01.4f s, " .
+					"Total Pollers: %s, " .
+					"Method: %s, " .
+					"Processes: %s, " .
+					"Threads: %s, " .
+					"Hosts: %s, " .
+					"Hosts/Process: %s",
+					round($end-$start,4),
+					$num_pollers,
+					$method,
+					$concurrent_processes,
+					$max_threads,
+					sizeof($all_polling_hosts),
+					$hosts_per_file),true,"SYSTEM");
+
+				db_execute("UPDATE poller SET run_state='Idle' WHERE active='on'");
+
+				break;
+			}
 		}else {
 			if (read_config_option("log_verbosity") >= POLLER_VERBOSITY_MEDIUM) {
-				print "Waiting on " . ($process_file_number - sizeof($polling_items)) . "/$process_file_number pollers.\n";
+				print "Waiting on " . ($process_file_number - sizeof($polling_items)) . "/$process_file_number Processes to Complete.\n";
 			}
 
-			process_poller_output($rrdtool_pipe);
+			/* process RRD output if any exists */
+			if ($poller_id == 0) {
+				process_poller_output($rrdtool_pipe);
+			}
 
 			/* end the process if the runtime exceeds MAX_POLLER_RUNTIME */
 			if (($start + MAX_POLLER_RUNTIME) < time()) {
-				rrd_close($rrdtool_pipe);
-				cacti_log("Maximum runtime of " . MAX_POLLER_RUNTIME . " seconds exceeded. Exiting.", true, "POLLER");
+				/* close rrdtool if poller is 0 */
+				if ($poller_id == 0) { rrd_close($rrdtool_pipe); }
+				cacti_log("Poller[$poller_id] ERROR: Maximum runtime of " . MAX_POLLER_RUNTIME . " seconds exceeded for Poller_ID " . $poller_id . " - Exiting.", true, "POLLER");
 				exit;
 			}
 
@@ -190,14 +266,14 @@ if (($num_polling_items > 0) && (read_config_option("poller_enabled") == "on")) 
 		}
 	}
 
-	rrd_close($rrdtool_pipe, $rrd_processes);
+	if ($poller_id == 0) { rrd_close($rrdtool_pipe, $rrd_processes); }
 
 	/* process poller commands */
 	$poller_commands = db_fetch_assoc("select
 		poller_command.action,
 		poller_command.command
 		from poller_command
-		where poller_command.poller_id=0");
+		where poller_command.poller_id=" . $poller_id);
 
 	$last_host_id = 0;
 	$first_host = true;
@@ -218,46 +294,56 @@ if (($num_polling_items > 0) && (read_config_option("poller_enabled") == "on")) 
 				}
 
 				if ($first_host) {
-					cacti_log("Host[$host_id] WARNING: Recache Event Detected for Host", true, "POLLER");
+					cacti_log("Poller[$poller_id] Host[$host_id] WARNING: Recache Event Detected for Host", true, "POLLER");
 				}
 
 				if (read_config_option("log_verbosity") == POLLER_VERBOSITY_DEBUG) {
-					cacti_log("Host[$host_id] RECACHE: Re-cache for Host, data query #$data_query_id", true, "POLLER");
+					cacti_log("Poller[$poller_id] Host[$host_id] RECACHE: Re-cache for Host, data query #$data_query_id", true, "POLLER");
 				}
 
 				run_data_query($host_id, $data_query_id);
 
 				if (read_config_option("log_verbosity") == POLLER_VERBOSITY_DEBUG) {
-					cacti_log("Host[$host_id] RECACHE: Re-cache successful.", true, "POLLER");
+					cacti_log("Poller[$poller_id] Host[$host_id] RECACHE: Re-cache successful.", true, "POLLER");
 				}
 			}
 		}
 
-		db_execute("delete from poller_command where poller_id=0");
+		db_execute("delete from poller_command where poller_id=" . $poller_id);
 
 		/* take time and log performance data */
 		list($micro,$seconds) = split(" ", microtime());
 		$recache = $seconds + $micro;
 
-		cacti_log(sprintf("STATS: " .
+		cacti_log(sprintf("Poller[$poller_id] STATS: " .
 			"Time: %01.4f s, " .
+			"Poller: %s",
 			"Hosts Recached: %s",
 			round($recache - $end,4),
+			$poller_id,
 			$recached_hosts),
 			true,"RECACHE");
 	}
 
-	/* graph export */
-	graph_export();
+	if ($poller_id == 0) {
+		/* graph export */
+		graph_export();
+
+		/* i don't know why we are doing this */
+		db_execute("truncate table poller_output");
+	}
 
 	if ($method == "cactid") {
 		chdir(read_config_option("path_webroot"));
 	}
-
-	db_execute("truncate table poller_output");
 }else{
-	print "There are no items in your poller cache or polling is disabled. Make sure you have at least one data source created. If you do, go to 'Utilities', and select 'Clear Poller Cache'.\n";
+	if ($poller_id == 0) {
+		cacti_log("Poller[$poller_id] ERROR: Either there are no pollers enabled, no items in your poller cache or polling is disabled. Make sure you have at least one data source created, your poller is active. If both are true, go to 'Utilities', and select 'Clear Poller Cache'.", true, "POLLER");
+	} else {
+   	db_execute("update poller set run_state = 'Complete' where poller_id=" . $poller_id);
+		cacti_log("Poller[$poller_id] WARNING: Poller had not items to process.", true, "POLLER");
+	}
 }
-// End Mainline Processing
+/* end mainline processing */
 
 ?>
